Refactored embedding.py so tables are created in db
